Refactor(DateTimeSection): extract pickers to separate component
