+++ packages/web/src/views/Forms/EventForm/DateTimeSection/DateTimeSection.tsx
@@ -1,25 +1,20 @@
 import React, { FC, SetStateAction } from "react";
 import dayjs from "dayjs";
 import customParseFormat from "dayjs/plugin/customParseFormat";
-import { Key } from "ts-key-enum";
 import { Categories_Event, Schema_Event } from "@core/types/event.types";
-import { MONTH_DAY_YEAR } from "@core/constants/date.constants";
-import { darken } from "@core/util/color.utils";
 import { SelectOption } from "@web/common/types/component.types";
 import { AlignItems } from "@web/components/Flex/styled";
 import { TimePicker } from "@web/components/TimePicker";
-import { DatePicker } from "@web/components/DatePicker";
 import {
-  dateIsValid,
   getTimeOptionByValue,
   getTimeOptions,
   mapToBackend,
-  shouldAdjustComplimentDate,
   shouldAdjustComplimentTime,
 } from "@web/common/utils/web.date.util";
 import { Option_Time } from "@web/common/types/util.types";
 
-import { StyledDateFlex, StyledDateTimeFlex, StyledTimeFlex } from "./styled";
+import { DatePickers } from "./DatePickers";
+import { StyledDateTimeFlex, StyledTimeFlex } from "./styled";
 
 dayjs.extend(customParseFormat);
 
@@ -64,30 +59,6 @@ export const DateTimeSection: FC<Props> = ({
 }) => {
   const timeOptions = getTimeOptions();
 
-  const adjustComplimentDateIfNeeded = (
-    changed: "start" | "end",
-    value: Date
-  ) => {
-    const start = changed === "start" ? value : selectedStartDate;
-    const end = changed === "end" ? value : selectedEndDate;
-
-    const { shouldAdjust, compliment } = shouldAdjustComplimentDate(changed, {
-      start,
-      end,
-    });
-
-    if (shouldAdjust) {
-      if (changed === "start") {
-        setSelectedEndDate(compliment);
-        return;
-      }
-
-      if (changed === "end") {
-        setSelectedStartDate(compliment);
-      }
-    }
-  };
-
   const adjustComplimentTimeIfNeeded = (
     changed: "start" | "end",
     value: string
@@ -125,84 +96,6 @@ export const DateTimeSection: FC<Props> = ({
     return defaultOption;
   };
 
-  const getDateFromInput = (val: string) => {
-    const date = dayjs(val, MONTH_DAY_YEAR).toDate();
-    return date;
-  };
-
-  const closeEndDatePicker = () => {
-    setIsEndDatePickerOpen(false);
-  };
-
-  const closeStartDatePicker = () => {
-    setIsStartDatePickerOpen(false);
-  };
-
-  const onPickerKeyDown = (
-    picker: "start" | "end",
-    e: React.KeyboardEvent<HTMLDivElement>
-  ) => {
-    switch (e.key) {
-      case Key.Backspace: {
-        e.stopPropagation();
-        break;
-      }
-      case Key.Enter: {
-        e.stopPropagation();
-        const input = e.target as HTMLInputElement;
-        const val = input.value;
-        const isInvalid = val !== undefined && !dateIsValid(val);
-
-        if (isInvalid) {
-          alert(`Sorry, IDK what to do with a ${picker} date of '${val}'
-          Make sure it's in '${MONTH_DAY_YEAR}' and try again`);
-          return;
-        }
-
-        const date = getDateFromInput(val);
-
-        if (picker === "start") {
-          onSelectStartDate(date);
-        }
-
-        if (picker === "end") {
-          onSelectEndDate(date);
-        }
-
-        break;
-      }
-      case Key.Escape: {
-        if (isStartDatePickerOpen) {
-          e.stopPropagation();
-          closeStartDatePicker();
-        }
-        if (isEndDatePickerOpen) {
-          e.stopPropagation();
-          closeEndDatePicker();
-        }
-        break;
-      }
-      case Key.Tab: {
-        if (isStartDatePickerOpen) {
-          setIsStartDatePickerOpen(false);
-        }
-        if (isEndDatePickerOpen) {
-          setIsEndDatePickerOpen(false);
-        }
-        break;
-      }
-      default: {
-        return;
-      }
-    }
-  };
-
-  const onSelectEndDate = (end: Date) => {
-    setSelectedEndDate(end);
-    setIsEndDatePickerOpen(false);
-    adjustComplimentDateIfNeeded("end", end);
-  };
-
   const onSelectEndTime = (option: SelectOption<string>) => {
     setEndTime(option);
     const correctedStart = adjustComplimentTimeIfNeeded("end", option.value);
@@ -224,12 +117,6 @@ export const DateTimeSection: FC<Props> = ({
 
       setEvent(_event);
     }
-  };
-
-  const onSelectStartDate = (start: Date) => {
-    setSelectedStartDate(start);
-    setIsStartDatePickerOpen(false);
-    adjustComplimentDateIfNeeded("start", start);
   };
 
   const onSelectStartTime = (option: SelectOption<string>) => {
@@ -255,62 +142,21 @@ export const DateTimeSection: FC<Props> = ({
     }
   };
 
-  const stopPropagation = (e: React.MouseEvent<HTMLDivElement>) => {
-    e.stopPropagation();
-  };
-
   return (
     <StyledDateTimeFlex alignItems={AlignItems.CENTER} role="tablist">
       {category === Categories_Event.ALLDAY && (
-        <>
-          <StyledDateFlex alignItems={AlignItems.CENTER}>
-            <div onMouseUp={stopPropagation} onMouseDown={stopPropagation}>
-              <DatePicker
-                bgColor={darken(bgColor, 15)}
-                calendarClassName="startDatePicker"
-                inputColor={inputColor}
-                isOpen={isStartDatePickerOpen}
-                onCalendarClose={closeStartDatePicker}
-                onCalendarOpen={() => {
-                  setIsStartDatePickerOpen(true);
-                }}
-                onChange={() => null}
-                onInputClick={() => {
-                  isEndDatePickerOpen && setIsEndDatePickerOpen(false);
-                  setIsStartDatePickerOpen(true);
-                }}
-                onKeyDown={(e) => onPickerKeyDown("start", e)}
-                onSelect={onSelectStartDate}
-                selected={selectedStartDate}
-                title="Pick Start Date"
-                view="grid"
-              />
-            </div>
-          </StyledDateFlex>
-
-          <StyledDateFlex alignItems={AlignItems.CENTER}>
-            <div onMouseUp={stopPropagation} onMouseDown={stopPropagation}>
-              <DatePicker
-                bgColor={darken(bgColor, 15)}
-                calendarClassName="endDatePicker"
-                inputColor={inputColor}
-                isOpen={isEndDatePickerOpen}
-                onCalendarClose={closeEndDatePicker}
-                onCalendarOpen={() => setIsEndDatePickerOpen(true)}
-                onChange={() => null}
-                onInputClick={() => {
-                  isStartDatePickerOpen && setIsStartDatePickerOpen(false);
-                  setIsEndDatePickerOpen(true);
-                }}
-                onKeyDown={(e) => onPickerKeyDown("end", e)}
-                onSelect={onSelectEndDate}
-                selected={selectedEndDate}
-                title="Pick End Date"
-                view="grid"
-              />
-            </div>
-          </StyledDateFlex>
-        </>
+        <DatePickers
+          bgColor={bgColor}
+          inputColor={inputColor}
+          isEndDatePickerOpen={isEndDatePickerOpen}
+          isStartDatePickerOpen={isStartDatePickerOpen}
+          selectedEndDate={selectedEndDate}
+          selectedStartDate={selectedStartDate}
+          setSelectedEndDate={setSelectedEndDate}
+          setSelectedStartDate={setSelectedStartDate}
+          setIsEndDatePickerOpen={setIsEndDatePickerOpen}
+          setIsStartDatePickerOpen={setIsStartDatePickerOpen}
+        />
       )}
 
       {category === Categories_Event.TIMED && (
